easytess-backend/requirements.txt
Flask
Flask-Cors
pytesseract
easyocr
opencv-python
Pillow
numpy
python-dotenv
pypdfium2
pyzbar
